update use-cases for delete and put posts with blog-id in params

+++ src/features/blogs/application/useCases/delete-post-from-blogs-in-params.use-case.ts
@@ -1,11 +1,14 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { BlogsRepositoryTO } from '../../infrastructure/blogs.repository.to';
 import { PostsRepositoryTO } from '../../../posts/infrastructure/posts.repository.to';
+import {UsersService} from "../../../users/application/users.service";
+import {UsersCheckHandler} from "../../../users/domain/users.check-handler";
 
 export class DeletePostWithBlogInParamsCommand {
   constructor(
     public postId: string,
-    public blogId: string
+    public blogId: string,
+    public bearerHeader?: string
   ) {
   }
 
@@ -16,13 +19,20 @@ export class DeletePostWithBlogInParamsUseCase
   implements ICommandHandler<DeletePostWithBlogInParamsCommand> {
   constructor(
     private readonly blogsRepository: BlogsRepositoryTO,
-    private readonly postsRepository: PostsRepositoryTO
+    private readonly postsRepository: PostsRepositoryTO,
+    private readonly usersService: UsersService,
+    private readonly usersCheckHandler: UsersCheckHandler
   ) {
   }
 
   async execute(command: DeletePostWithBlogInParamsCommand) {
     const findedBlog = await this.blogsRepository.findBlogById(command.blogId)
-    const deletePost = await this.postsRepository.deletePostFromBlogsUri(command.postId, command.blogId)
-    return deletePost
+    if (!command.bearerHeader) {
+      return await this.postsRepository.deletePostFromBlogsUri(command.postId, command.blogId)
+    }
+    const user = await this.usersService.getUserByAuthToken(command.bearerHeader);
+    if (this.usersCheckHandler.checkIsOwner(Number(findedBlog.userId), Number(user.id))) {
+      return await this.postsRepository.deletePostFromBlogsUri(command.postId, command.blogId)
+    }
   }
 }
